# planning-service/openapi.json
{
  "components": {
    "schemas": {
...
      "AirportPin": {
        "description": "A landing field the map can open a card for: the ident pilots use,\nwhere it is, and its size -- the bigger ones are kept when a wide\nview holds more than the map asks for. `flight_category` is the\nfield's METAR's (VFR, MVFR, IFR, LIFR), for the map's weather chip:\nNone where it has no reporting station, or the weather could not be\nasked.",
        "properties": {
          "flight_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Category"
          },
          "ident": {
            "title": "Ident",
            "type": "string"
          },
          "kind": {
            "enum": [
              "large",
              "medium",
              "small",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "ident",
          "name",
          "lat",
          "lon",
          "kind"
        ],
        "title": "AirportPin",
        "type": "object"
      },
      "AirportPlace": {
        "description": "One airport the way the map's card shows it: its name and place,\nthe class of the airspace over it, whether it has a tower, its\nrunways and radio, and the weather there now. `metar` is None for a\nfield with no reporting station, and `weather_unavailable` says the\nweather service could not be asked at all -- not the same thing.",
        "properties": {
          "airport_diagram_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Airport Diagram Url"
          },
          "airspace_class": {
            "anyOf": [
              {
                "enum": [
                  "B",
                  "C",
                  "D",
                  "E",
                  "G"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Airspace Class"
          },
          "chart_supplement_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Supplement Url"
          },
          "elevation_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation Ft"
          },
          "frequencies": {
            "items": {
              "$ref": "#/components/schemas/Frequency"
            },
            "title": "Frequencies",
            "type": "array"
          },
          "ident": {
            "title": "Ident",
            "type": "string"
          },
          "kind": {
            "enum": [
              "large",
              "medium",
              "small",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lighting": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Lighting",
            "type": "array"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "metar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metar"
              },
              {
                "type": "null"
              }
            ]
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "radio_notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Radio Notes",
            "type": "array"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "runways": {
            "items": {
              "$ref": "#/components/schemas/Runway"
            },
            "title": "Runways",
            "type": "array"
          },
          "standard_keying": {
            "default": false,
            "title": "Standard Keying",
            "type": "boolean"
          },
          "towered": {
            "title": "Towered",
            "type": "boolean"
          },
          "weather_unavailable": {
            "default": false,
            "title": "Weather Unavailable",
            "type": "boolean"
          }
        },
        "required": [
          "ident",
          "name",
          "lat",
          "lon",
          "kind",
          "towered",
          "runways",
          "frequencies"
        ],
        "title": "AirportPlace",
        "type": "object"
      },
...
      "AirportsInView": {
        "properties": {
          "airports": {
            "items": {
              "$ref": "#/components/schemas/AirportPin"
            },
            "title": "Airports",
            "type": "array"
          }
        },
        "required": [
          "airports"
        ],
        "title": "AirportsInView",
        "type": "object"
      },
...
      "Frequency": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frequency_mhz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency Mhz"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "Frequency",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Metar": {
        "properties": {
          "altimeter_in_hg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altimeter In Hg"
          },
          "ceiling_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ceiling Ft"
          },
          "dewpoint_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dewpoint C"
          },
          "flight_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Category"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "temp_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp C"
          },
          "visibility_sm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility Sm"
          },
          "wind_dir_true_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Dir True Deg"
          },
          "wind_gust_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Gust Kt"
          },
          "wind_speed_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Kt"
          }
        },
        "title": "Metar",
        "type": "object"
      },
...
      "NearestAirport": {
        "description": "A field near a position: how far and which way (true), its town,\nelevation and longest open runway.",
        "properties": {
          "bearing_deg": {
            "title": "Bearing Deg",
            "type": "number"
          },
          "distance_nm": {
            "title": "Distance Nm",
            "type": "number"
          },
          "elevation_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation Ft"
          },
          "flight_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Category"
          },
          "ident": {
            "title": "Ident",
            "type": "string"
          },
          "kind": {
            "enum": [
              "large",
              "medium",
              "small",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "longest_runway_ft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longest Runway Ft"
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "ident",
          "name",
          "lat",
          "lon",
          "kind",
          "distance_nm",
          "bearing_deg"
        ],
        "title": "NearestAirport",
        "type": "object"
      },
      "NearestAirports": {
        "properties": {
          "airports": {
            "items": {
              "$ref": "#/components/schemas/NearestAirport"
            },
            "title": "Airports",
            "type": "array"
          }
        },
        "required": [
          "airports"
        ],
        "title": "NearestAirports",
        "type": "object"
      },
...
      "Runway": {
        "properties": {
          "closed": {
            "title": "Closed",
            "type": "boolean"
          },
          "ends": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends"
          },
          "length_ft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Ft"
          },
          "lighted": {
            "title": "Lighted",
            "type": "boolean"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "width_ft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Ft"
          },
          "wind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunwayWind"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ends",
          "length_ft",
          "width_ft",
          "surface",
          "lighted",
          "closed"
        ],
        "title": "Runway",
        "type": "object"
      },
      "RunwayWind": {
        "description": "The reported wind on the end of a runway it favours, the one with\nthe most headwind (vfr.runway_wind): the headwind negative for a\ntailwind, the crosswind positive from the right, and the crosswind\nin the gusts where the report has them.",
        "properties": {
          "crosswind_kt": {
            "title": "Crosswind Kt",
            "type": "number"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "gust_crosswind_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gust Crosswind Kt"
          },
          "headwind_kt": {
            "title": "Headwind Kt",
            "type": "number"
          }
        },
        "required": [
          "end",
          "headwind_kt",
          "crosswind_kt"
        ],
        "title": "RunwayWind",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WaypointPin": {
        "description": "A VFR waypoint on the sectional (VPBNG), for the map's diamond: a\ntap names it, and with a route open puts it in the stops.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ident": {
            "title": "Ident",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "ident",
          "lat",
          "lon"
        ],
        "title": "WaypointPin",
        "type": "object"
      },
      "WaypointsInView": {
        "properties": {
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/WaypointPin"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "waypoints"
        ],
        "title": "WaypointsInView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Wingtip Maps planner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/airport/{ident}": {
      "get": {
        "description": "One US airport's card, by any ident it goes by (C81, KC81, KDLH).\n\n`airspace_class` is the class of the airspace at the surface at the\nfield: B, C or D, E in a Class E surface area, G where the ground is\nuncontrolled (vfr.airspace.surface_class_at).\n`towered` is whether it lists a tower frequency. A field with no\nreporting station has no `metar`; `weather_unavailable` is set only\nwhen the weather service could not be asked.",
        "operationId": "airport_place_api_airport__ident__get",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportPlace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Airport Place"
      }
    },
    "/api/airports/in-view": {
      "get": {
        "description": "The landing fields inside the map's view, the biggest first --\nat most `limit` of them, since a whole state holds thousands and the\nmap only lays a tap target over each -- with each one's flight\ncategory from its METAR, for a weather chip on the ones that report.\nThe METARs are the national cache already in memory; with the\nweather service out the fields come back without, as fields with no\nstation do. With `reporting`, only the fields that report, before\nthe limit: what the map asks once for a whole route's box, so its\nchips are there before the pilot zooms in on any of it.",
        "operationId": "airports_in_view_api_airports_in_view_get",
        "parameters": [
          {
            "in": "query",
            "name": "south",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "South",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "west",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "West",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "north",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "North",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "east",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "East",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reporting",
            "required": false,
            "schema": {
              "default": false,
              "title": "Reporting",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportsInView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Airports In View"
      }
    },
    "/api/airports/nearest": {
      "get": {
        "description": "The landing fields nearest a position -- own ship's, for the map's\nNearest -- the nearest first: how far and which way, each one's\nlongest open runway and its flight category where it reports.",
        "operationId": "nearest_airports_api_airports_nearest_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearestAirports"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nearest Airports"
      }
    },
...
    "/api/waypoints/in-view": {
      "get": {
        "description": "The VFR waypoints inside the map's view (vfr.fixes), at most\n`limit`: the magenta flags the sectional prints round busy airspace,\nfor the map to mark each with a diamond a pilot can tap and route\nthrough. GPS waypoints and the rest of NASR's fixes are left out:\nthey are not on a VFR chart.",
        "operationId": "waypoints_in_view_api_waypoints_in_view_get",
        "parameters": [
          {
            "in": "query",
            "name": "south",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "South",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "west",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "West",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "north",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "North",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "east",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "East",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaypointsInView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Waypoints In View"
      }
    }
  }
}